# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nextcloud WeKnora development API
  version: 0.4.10
  description: Local integration contract. Machine calls require a binding-scoped Bearer key and signed headers specified in docs/machine-request-signing.md. Multi-page manifests use a short-lived stored list, but source changes can still invalidate it. The source authorization API returns the current file ETag for a guarded read. The paired WeKnora patch covers tested read paths; independent user identity proof, enterprise permission and publication acceptance remain outstanding.
servers:
  - url: /index.php/apps/integration_weknora/api/v1
security:
  - serviceToken: []
    machineKeyId: []
    machineTimestamp: []
    machineNonce: []
    machineSignature: []
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: Secret token whose SHA-256 digest is also the binary HMAC key.
    machineKeyId:
      type: apiKey
      in: header
      name: X-WeKnora-Key-Id
    machineTimestamp:
      type: apiKey
      in: header
      name: X-WeKnora-Timestamp
    machineNonce:
      type: apiKey
      in: header
      name: X-WeKnora-Nonce
    machineSignature:
      type: apiKey
      in: header
      name: X-WeKnora-Signature
    nextcloudSession:
      type: apiKey
      in: header
      name: Cookie
      description: Authenticated Nextcloud session. Administrator routes additionally require system administrator privileges.
    csrfToken:
      type: apiKey
      in: header
      name: requesttoken
      description: Nextcloud CSRF token for mutating requests.
  schemas:
    Binding:
      type: object
      required: [id, name, root_file_id]
      properties:
        id:
          type: string
        name:
          type: string
        root_file_id:
          type: integer
...
paths:
...
  /bindings:
    get:
      operationId: listBindings
      description: Returns only the one binding attached to the authenticated machine key.
      responses:
        '200':
          description: Published folder bindings
          content:
            application/json:
              schema:
                type: object
                required: [bindings]
                properties:
                  bindings:
                    type: array
                    items: {$ref: '#/components/schemas/Binding'}
        '401': {description: Invalid service token}
...